refactor(activities): who was on a message gets its own file
emailpresentation.go crossed the 500-line cap. The recipients — who was on
it, and which of them the row names — are a concept of their own, answered
from activity_participant rows under the caller's row scope rather than from
anything the presentation assembles.
